Fix initial communication detection
We were listening for a Login event, but that only shows that someone *else* has logged in -- it's not a
reliable indicator. Instead, we should watch for the Network event, which shows we're connected to a DVR, and
a message with multiple VideoInput events, indicating that we've received the full list of video channels.