Add support for sensors (alarm contacts)
The Sensor event includes a bitmask of asserted alarms. Based on that,
create binary_sensor entities for sensors up to the highest number that
was asserted. Hopefully this will give true dynamic detection, as I
don't see another way to determine which ones actually exist, let alone
are exposed to the outside.