Tighten up message processing more
The goal is to make the whole message processing pipeline robust. I've
seen occasional glitches where we lost sync, even though the DVR didn't
do anything funny. Rework the handlers so we use common code, and handle
the irregularities in that code, not some sequence-specific place.

Once we've started the real message stream, the MIME parts are the outer
container, so make sure we really treat the boundary as the ultimate
separator between messages. If we lose sync, we should look for the next
boundary, not the next 4-byte prefix.

Also, we shouldn't have a bunch of special handlers at the beginning of
the connection. An emsg is an emesg. Just sometimes, they define a
boundary, and sometimes they contain data. The first one might contain
the leftover HTTP status message from the connection upgrade to
websocket, but we just handle that when we see it.

Touch up the log messaging a little bit so we're not dumping a ton of
noise when something goes wrong.