Rework authentication and boundary detection
Even though the web client does it, I don't think we need to send auth a
second time in response to wsev. Just treat wsev as a form of
acknowledgement.

Rework the boundary detection a little bit more, after stressing it a
few times. We need to be very clear at what we consume as an emsg, and
understand that the first one both defines and ends with a boundary.